[FIX] LikeServiceImpl 분리. 주석 추가.
